KP-7481 Remove assert from production code
The assert will not produce a specific catchable exception nor
informative error message, so it doesn't seem to add much value beyond
the TypeError we will get on the next line if the Metax ID is not found
(which should not happen in the first place). As it doesn't bring much
value, it was removed.