KP-7947 Add helper for directly retrieving corpora only
Previously filtering out non-corpora or pidless records had to be done
at the receiving end, now PMH_API can provide pre-filtered data for easy
use. This means that the same filtering code does not need to be
repeated in different locations.

Also renamed the corpora dictifying function to emphasize that it does
not handle all metadata records, just the corpora.