KP-7947 Use Click for command line arguments
It would not have been strictly necessary to argumentify the log file:
it could still have been left hard-coded, but keeping it hard-coded
while adding the configuration file would have seemed silly.

As we will need some way to handle input files when the configuration
file is added, the overall adoption of click (or something else, but
click is nice as it takes care of file existence checking etc) will be
necessary anyway.