KP-7947 Add API token to Metax queries, provided via yaml config
The Metax test environment has been updated so that it now requires
authorization when adding or editing records. This change allows us to
provide it.

The token is read from a configuration file that the user specifies when
invoking the CLI. The goal is that if there's a problem with the
configuration (e.g. missing values or wrong format), the CLI will always
provide the user with a helpful error message.