KP-7973 Explicitly differentiate datetime vs date in metadata class, drop secfrac
`_get_date` returning a datetime was confusing, so renamed it to
`_get_datetime` in preparation for the day when we actually use fields
that require a date instead of datetime.

Also dropped the secfrac from the timestamp: Metax does not require it,
it isn't present in the default datetime, and it does not make sense to
specify fractions of a second for a field whose original data is at
granularity of one day.