KP-7973 Use `created` for metadata creation timestamp
According to Metax docs, `issued` is "Date of formal issuance (e.g.,
publication) of the resource", which is not what the
`metadataCreationDate` from Metashare depicts. Not providing the
`created` attribute but providing `modified` also causes errors during
metadata validation: the default value for `created` is the time of
creation for the record, which means that `created` would be later than
`modified`.

Now we don't provide `issued` at all (it is not consistently available
in Metashare metadata and the field can be empty), but do use `created`
and `modified` in a consistent way.