KP-7973 Prevent sending empty `DatasetLicense` dicts to Metax
If we have licenses that are not present in `license_mappings` (in
practice because therey are not present in suomi.fi reference catalog),
they won't get a custom URL either. This can result in an empty license
dict with neither `url` nor `custom_url`, which is not allowed in Metax.
To avoid this, empty license dicts are simply skipped.