KP-7973 Improve missing license URL handling
It might be that none of the licenses listed in the Metashare metadata
have a mapped counterpart in suomi.fi license codes, which would result
in an empty license list despite there having been license elements.
Now we add the "other" license in that case.

Also simplified the license element extraction so that it always returns
a list, even if it happens to be empty (previously `None` was returned
instead of empty list).