KP-7427 Merge same actor in multiple roles into one actor with multiple roles
The old implementation created multiple Actors if the same person held
multiple roles. Now multiple roles can be represented for the same
actor.

This means also that the `person: None` entries will be merged,
which is silly. Nothing was done to alleviate the silliness, because
these actors will get content when KP-7425 is implemented.

Making the roles set into a roles list is an operation that does not
guarantee to produce the same order in the resulting list. This is
problematic when testing, so the roles are now sorted in the metaxified
output.