KP-7425 Add non-URI organization data and handle translations
This is a bit ugly combination of two changes in functionality:
  1. Changed the XML parsing so that multiple language variants of the
     same field are all processed individually
  2. Sending a full metadata dict for organizations that do not have a
     suomi.fi URI.

In theory these are separate concerns, but feature 1 had to be
implemented so that different languages given for organization names can
be handled in 2. Because earlier we ignored the `lang` label, we ended
up overwriting a field if another language variant was encountered. Now
all the language variants are parsed and stored separately, and one (for
person name) or all (for organization names) are used in the Metax
metadata. Current setting is to only use Finnish, English or "undefined"
fields (in that order of preference if only one is desired).

For number 2, we simply populate the `pref_label` (in all available
supported languages), `homepage` (as a dict that only contains the page
URL under "identifier" key, "title" is not provided) and `email`.