KP-8714 Metadata parsing for CMDI format from COMEDI
The test data has been replaced with roughly corresponding data from
COMEDI, and the metadata parser has been adjusted so that the correct
fields are extracted from the test data set records. Despite the tests
for metadata parser passing, the code is not always pretty nor is it
even guaranteed to work with a wider selection of records from the full
data set, so further testing is needed.