KP-8714 Get rid of Metashare in names and comments
As an added bonus, the test data files that deliver the full OAI-PMH
response are prefixed with `comedi_list_records_` to separate them from
the test files that only contain a single `record`. The former are
needed when testing involves the OAI-PMH requests whereas the latter are
used when testing just the metadata parsing.