KP-8714 Make publisher parsing work for persons
CMDI has separate distributionRigtsHolderPerson and
distributionRightsHolderOrganization elements, which is a bit tricky for
the current parsing. At the moment, sole organizations are not added as
actors at all though:
```
$ git show b815790f
commit b815790f952b20950564a768e5cf3157c0724cf3
Author: Anni Järvenpää <anni.jarvenpaa@csc.fi>
Date:   Tue Feb 27 11:37:54 2024 +0000

    KP-7427 Skip empty actors

    Because the organizations have not been implemented yet, some of the
    actors had just a role and an empty person dict. Metax does not like
    this, so those are skipped altogether now.
```

Thus it's less bad to read the person element than to read the
organization element. This needs to be made to work with both
organizations and persons though if (when?) we remove the person data
requirement is lifted.