KP-8714 Include organization-only actors
On the contrary to the previous implementation, Metax requires each
actor to have at least URL for the organization ("Each actor has an
organization and optionally a person"
https://metax-service.fd-staging.csc.fi/v3/docs/user-guide/datasets-api/#actors).
At this stage, support was added to actors without person info, but
nothing has (yet) been done to the fact that we will still offer Metax
actors without an organization (common e.g. for metadata creators).