KP-9470 Always fix the number of publishers to one
Metax requires there to be exactly one publisher for each dataset. This
is not true in Comedi source metadata: there might be more than one or
there might be none. The situation where there are none must be fixed in
the source data (that is likely not true, as the data has clearly been
published so someone should have right to do that). When there are more
than one though, we purge the original publishers and add a fake
organization that lets the user know that they need to see the original
metadata for accurate publisher information.