KP-8261 Drop urn.fi from persistent identifiers uploaded to Metax
Fairdata services seem to expect only the urn:nbn:fi:lb-01234 part as
persistent_identifier (or preferred_identifier in older API versions).
Failure to comply with this expectation leads both to the identifier
field in Etsin not functioning as a clickable button and our records
from this new bridge and the old harvests by Metax team not matching up,
resulting in double entries for each dataset if not fixed.