KP-9691 Default to not automatically deleting records
To play it safe, we want our cronjobs to only report the need for
deleting records instead of actually deleting the records. These changes
make the CLI tool default to just listing the records to be deleted. If
the list looks sensible, the tool can still carry out the deletion
operations if explicitly told to do so by providing the
`--automatic-delete` flag.