KP-9760 Switch from fd-rework to demo API for testing
Fairdata team is deprecating the old fd-rework.csc.fi, so we must also
start using metax.demo.fairdata.fi, so it seems reasonable to reflect
this in the configuration example too to avoid possible confusion in the
future.

Adjusting the tests would not be strictly necessary as there the URL
does not need to match that of the real service, but it's a small job
and leaving them as the old URL might be a potential source of confusion
too.