KP-9717 Determine creator based on resourceCreator
The old approach matched the spec created when initiating the project,
but it is not what we want: creator in Metax should be the creator of
the dataset, not creator of the associated metadata. The change also
required adjusting many of the test files due to resourceCreators not
having been present earlier.