KP-8075 Add the full payload to the log on Metax errors
This has potential to make the logs overly verbose, but on the other
hand it will be a tremendous aid in debugging when we need to get back
to problems that have occurred in a scheduled run. Further, these errors
should be rare enough to not pose a problem to the log volume.