KP-10725 Add CLI tool for listing Metax-incompatibilities
The tool will detect both problems we have an explicit check for and
ones we don't, but this requires actually submitting the records to a
Metax instance. If this is a demo instance, this is not problematic, but
to verify that the tool is not accidentally used against production,
there is an explicit verification step before proceeding.