KP-10529 Add option to save harvested records locally
The records are saved in individual XML files under a given directory.
To enable parsing the boolean value for whether to save the records, the
parsing of the configuration file has now been boosted to use
`SafeLoader`.

We want to save all records, not only the published corpora, so the
saving is done in a separate loop instead of as part of the normal
bridge loop. The records whose saving has failed are also reported
separately.

If local backup is enabled, all records are always redownloaded instead
of only the records changed since the last run. This ensures that runnin
the bridge once without backing up enabled will not leave gaps in the
backup.

The tests must now account for the difference between requests for
published or unpublished records towards COMEDI. Because most of the
tests don't create backups, most COMEDI request fixtures can stay
ignorant of different statuses, but the ones involved in testing backing
up the two types of records must now have the status.