KP-10529 Delete old records from backup before fetching new ones
We want our backups to be restorable to COMEDI to reach the situation at
the time when the backup was created. This means, that if a record is
removed from COMEDI, it must also not be present in our backup. The
trivial way to achieve this is deleting all old files before fetching the
new backed up batch. As an extra safeguard, we check the file names
before deleting, and halt the program if we encounter files that do not
match the generated backup file name pattern.